Initial commit of stats processing scripts.
These scripts process the stats files from TBLoader into the PostgreSQL database.

The scripts run from withing a Dropbox directory, and are in git for versioning
and for safekeeping. The install.sh script copies the scripts on top of an
existing AWS-LB directory, so that saved log files, notes, or examples will
not be touched.