New production, deployment, and usage statistics.
New stats based on an initial "guess" at what will be most informative for
partners. All reports produced as .csv files. Reports by project include:
- production by content update
-- by update and package
-- by update, package, and category
- deployments by content update and package
- usage by content update
-- by update and category
-- by update, package, and category
-- by update, package, category, and message

There are two global "dashboard" reports, with usage and deployment summaries
for the last (up to) 4 content updates for every project.

The new reports are generated by a scheme in which some "helper" queries are
prepended to some very minimal queries, and wrapped in sql to copy to a .csv
file. This is a bit of an experiment, to see how well it deals with updating
the queries.