Fixes to the reports scripts.
Started using a file for the filter, because that's easier for the script.

It turns out that avoiding the symlink "Dropbox -> Dropbox\ \(Literacy\ Bridge\)" is a bad idea,
because of the spaces in the filename, which plays badly with the reports shell script.

So, this change returns to using the symlink.

Dropbox: fabulous for knowledge workers, not so much for production workflows.