Added script used to correct langauge code.
Script to do a lot of the work of correcting a language code. In this
case, ssl1 was used incorrectly, rather than the correct sil, to
represent Tumu Sisaala.

This script updates the database tables and metadata .csv files,
and renames ...TB-Loaders/communities/XYZ/languages/ssl1 and
.../XYZ/system/ssl1.grp with "sil".

The ACM database is handled separately. (It's not very amenable to
command line editing.)