refactor!: Eliminate Dropbox in production.
Upcoming changes to Dropbox will change the on-disk directory structure.
It is not clear how that will even work on Linux, which is not
officially supported by Dropbox. We have previously moved all ACM
content data from Dropbox, so that no Amplio software user needs
Dropbox, but the statistics processing still used Dropbox to archive
processed data, and to distribute the scripts and applications used to
process the statistics and UF.

Changes:
- Use AWS S3 as the archival location for processed data, rather than
  Dropbox. Dropbox/collected-data-processed -> s3://acm-stats/processed-data
- Use AWS S3 as the deployment location for scripts and applications,
  rather than Dropbox. Dropbox/AWS-LB -> s3://acm-stats/AWS-LB
- Update various scripts to use the new locations.
- Remove obsolete scripts and applications that were used when Dropbox
  was part of the production workflow.

BREAKING CHANGE: The processed statistics data no longer flow into
Dropbox/collected-data-processed/YYY, but now are copied into
s3://acm-stats/processed-data/YYYY. While this is, strictly speaking, a
breaking change, only a single developer (this committer) is a consumer
of that data.